+++ core/core.js
@@ -9,7 +9,7 @@ export { renderAnnotationsSection, mapZoteroAnnotation } from "../src/annotation
 export { syncBlocks, makeBlock, parseBlocks, migrateLegacyAnnotations } from "../src/blocks.js";
 export { DEFAULT_FORMATS } from "../src/formats.js";
 export { parseTemplateFile, templateKind, templateUserOwnedKeys } from "../src/templates.js";
-export { parseManifest, hasManifest, applyManifest, setManifestEntry, removeManifestEntry, buildManifestFromScaffold, writeManifest, MANIFEST_KEY, getTagField, setTagField } from "../src/manifest.js";
+export { parseManifest, hasManifest, applyManifest, setManifestEntry, removeManifestEntry, buildManifestFromScaffold, writeManifest, MANIFEST_KEY, getTagField, setTagField, getAttachmentFolder, setAttachmentFolder } from "../src/manifest.js";
 export { frontmatterList, cleanTag, tagSyncPlan } from "../src/tagsync.js";
 export { findMarkerRanges, rangeRevealed } from "../src/markers.js";
 export { findFrontmatterRange, findHeadingRanges, findLinkRanges, findEmphasisRanges } from "../src/preview.js";